app/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==2.1.1
packaging==24.1
pandas==2.2.3
plotly==5.24.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.35
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.4
